# cmd/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  client.ReceptionistList:
    properties:
      status:
        description: 接待人员的接待状态。0:接待中,1:停止接待。第三方应用需具有“管理帐号、分配会话和收发消息”权限才可获取
        type: integer
      userid:
        description: 接待人员的userid。第三方应用获取到的为密文userid，即open_userid
        type: string
    type: object
...
  dto.Response:
    properties:
      code:
        type: integer
      data: {}
      message:
        type: string
    type: object
  dto.SignatureOptions:
    properties:
      echostr:
        type: string
      msg_signature:
        type: string
      nonce:
        type: string
      timestamp:
        type: string
    type: object
...
  request.ReceptionistOptions:
    properties:
      userid_list:
        items:
          type: string
        type: array
    type: object
  request.Staff:
    properties:
      email:
        type: string
      number:
        type: string
      policy_list:
        items:
          type: string
        type: array
      staffid:
        type: string
      staffname:
        type: string
    type: object
  request.WecomConfigApp:
    properties:
      agent_id:
        type: string
      corp_id:
        type: string
      encoding_aes_key:
        type: string
      secret:
        type: string
      token:
        type: string
      type:
        type: string
    type: object
host: localhost
info:
  contact: {}
  description: evobot
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: EvoBot
  version: "1.0"
paths:
...
  /staff:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all staff members
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: List Staff
      tags:
      - staff
    post:
      consumes:
      - application/json
      description: Add a new staff member
      parameters:
      - description: Staff Information
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.Staff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Add Staff
      tags:
      - staff
  /staff/{uuid}:
    delete:
      consumes:
      - application/json
      description: Delete the staff member by UUID
      parameters:
      - description: Staff UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Delete Staff
      tags:
      - staff
    get:
      consumes:
      - application/json
      description: Get the staff information by UUID
      parameters:
      - description: Staff UUID
        in: path
        name: uuid
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get Staff
      tags:
      - staff
    patch:
      consumes:
      - application/json
      description: Update the staff member by UUID
      parameters:
      - description: Staff UUID
        in: path
        name: uuid
        required: true
        type: string
      - description: Staff Information
        in: body
        name: req
        required: true
        schema:
          $ref: '#/definitions/request.Staff'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update Staff
      tags:
      - staff
...
  /wecom/callback:
    get:
      consumes:
      - application/json
      description: Verifies the callback URL for WeCom (Enterprise WeChat). This endpoint
        is used to handle URL verification for WeCom's callback interface.
      parameters:
      - description: Wecom SignatureOptions data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/dto.SignatureOptions'
      produces:
      - application/json
      responses:
        "200":
          description: The echo string returned upon successful verification
          schema:
            type: string
      summary: WeCom Callback URL Verification
      tags:
      - wecom_callback
    post:
      consumes:
      - application/json
      description: Handles callback messages from WeCom (Enterprise WeChat). This
        endpoint receives and parses the callback data sent by WeCom.
      produces:
      - application/json
      responses:
        "200":
          description: Parsed WeCom callback message
          schema:
            type: string
      summary: WeCom Callback Message Handling
      tags:
      - wecom_callback
...
  /wecom/config/{uuid}:
    get:
      consumes:
      - application/json
      description: Retrieve the current Wecom configuration
      produces:
      - application/json
      responses:
        "200":
          description: Current Wecom configuration
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Get Wecom configuration
      tags:
      - wecom_config
    patch:
      consumes:
      - application/json
      description: Update the Wecom configuration with the provided data
      parameters:
      - description: Wecom configuration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.WecomConfigApp'
      produces:
      - application/json
      responses:
        "200":
          description: Success response
          schema:
            $ref: '#/definitions/dto.Response'
      summary: Update Wecom configuration
      tags:
      - wecom_config
  /wecom/receptionist/{kfid}:
    delete:
      consumes:
      - application/json
      description: receptionist delete
      parameters:
      - description: Wecom configuration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReceptionistOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: WeCom receptionist delete
      tags:
      - wecom_receptionist
    get:
      consumes:
      - application/json
      description: receptionist list
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/client.ReceptionistList'
            type: array
      summary: WeCom receptionist list
      tags:
      - wecom_receptionist
    post:
      consumes:
      - application/json
      description: receptionist add
      parameters:
      - description: Wecom configuration data
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.ReceptionistOptions'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      summary: WeCom receptionist add
      tags:
      - wecom_receptionist
swagger: "2.0"